[Gaarahan] feature: modify UserDto to fit new business and fix the related problem

+++ src/main/java/com/thoughtworks/rslist/api/UserController.java
@@ -3,6 +3,7 @@ package com.thoughtworks.rslist.api;
 import com.thoughtworks.rslist.domain.User;
 import com.thoughtworks.rslist.dto.UserDto;
 import com.thoughtworks.rslist.exception.InvalidIndexException;
+import com.thoughtworks.rslist.exception.InvalidUserIdException;
 import com.thoughtworks.rslist.repository.UserRepository;
 import org.modelmapper.ModelMapper;
 import org.springframework.http.ResponseEntity;
@@ -12,6 +13,7 @@ import javax.transaction.Transactional;
 import javax.validation.Valid;
 import java.util.List;
 import java.util.Optional;
+import java.util.stream.Collectors;
 
 /**
  * @author gaarahan
@@ -19,7 +21,7 @@ import java.util.Optional;
 @RestController
 public class UserController {
   private final UserRepository userRepository;
-  ModelMapper mapper = new ModelMapper();
+  ModelMapper modelMapper = new ModelMapper();
 
   public UserController(UserRepository userRepository) {
     this.userRepository = userRepository;
@@ -27,29 +29,42 @@ public class UserController {
 
   @PostMapping("/users")
   public void register(@RequestBody @Valid User user) {
-    UserDto newUser = mapper.map(user, UserDto.class);
+    UserDto newUser = modelMapper.map(user, UserDto.class);
     this.userRepository.save(newUser);
   }
 
   @GetMapping("/users")
-  public ResponseEntity<List<UserDto>> getList() {
-    return ResponseEntity.ok(this.userRepository.findAll());
+  public ResponseEntity<List<User>> getList() {
+    List<User> allUser = this.userRepository.findAll().stream()
+        .map(userDto -> modelMapper.map(userDto, User.class))
+        .collect(Collectors.toList());
+    return ResponseEntity.ok(allUser);
   }
 
   @GetMapping("/users/{id}")
-  public ResponseEntity<Optional<UserDto>> getUserById(@PathVariable Integer id) {
+  public ResponseEntity<User> getUserById(@PathVariable Integer id) {
     if (id == null || id < 0) {
-      throw new InvalidIndexException();
+      throw new InvalidUserIdException();
     }
-    return ResponseEntity.ok(this.userRepository.findById(id));
+    Optional<UserDto> userDto = this.userRepository.findById(id);
+    if (!userDto.isPresent()) {
+      throw new InvalidUserIdException();
+    }
+    return ResponseEntity.ok(modelMapper.map(userDto.get(), User.class));
   }
 
   @Transactional
-  @GetMapping("/users/delete/{id}")
+  @DeleteMapping("/users/{id}")
   public void deleteUserById(@PathVariable Integer id) {
     if (id == null || id < 0) {
       throw new InvalidIndexException();
     }
+
+    Optional<UserDto> userDto = this.userRepository.findById(id);
+    if (!userDto.isPresent()) {
+      throw new InvalidUserIdException();
+    }
+
     this.userRepository.deleteById(id);
   }
 }
